Added an authorization mechanizm to kay.generics package.

+++ kay/generics/__init__.py
@@ -12,7 +12,9 @@ from string import Template
 from werkzeug.routing import (
   Rule, RuleTemplate, EndpointPrefix, Submount,
 )
-from werkzeug.exceptions import NotFound
+from werkzeug.exceptions import (
+  NotFound, Forbidden
+)
 from werkzeug import (
   Response, redirect
 )
@@ -23,6 +25,7 @@ from kay.utils import (
 from kay.utils.flash import (
   set_flash, get_flash
 )
+from kay.exceptions import NotAuthorized
 from kay.i18n import gettext as _
 from kay.routing import ViewGroup
 
@@ -34,23 +37,38 @@ endpoints = {
   'delete': "delete_$model",
 }
 
+OPE_LIST = 'list'
+OPE_SHOW = 'show'
+OPE_CREATE = 'create'
+OPE_UPDATE = 'update'
+OPE_DELETE = 'delete'
+
+def only_owner_can_write(self, request, operation, obj=None):
+  if operation == OPE_CREATE:
+    if request.user.is_anonymous():
+      raise NotAuthorized()
+  elif operation == OPE_UPDATE or operation == OPE_DELETE:
+    owner = getattr(obj, self.owner_attr)
+    if owner != request.user:
+      raise NotAuthorized()
 
 class CRUDViewGroup(ViewGroup):
   entities_per_page = 20
   templates = {
-    'list': '_internal/general_list.html',
-    'show': '_internal/general_show.html',
-    'update': '_internal/general_update.html',
+    OPE_LIST: '_internal/general_list.html',
+    OPE_SHOW: '_internal/general_show.html',
+    OPE_UPDATE: '_internal/general_update.html',
   }
   forms = {}
   form = None
+  owner_attr = None
   rule_template = RuleTemplate([
-    Rule('/$model/list', endpoint=endpoints['list']),
-    Rule('/$model/list/<cursor>', endpoint=endpoints['list']),
-    Rule('/$model/show/<key>', endpoint=endpoints['show']),
-    Rule('/$model/create', endpoint=endpoints['create']),
-    Rule('/$model/update/<key>', endpoint=endpoints['update']),
-    Rule('/$model/delete/<key>', endpoint=endpoints['delete']),
+    Rule('/$model/list', endpoint=endpoints[OPE_LIST]),
+    Rule('/$model/list/<cursor>', endpoint=endpoints[OPE_LIST]),
+    Rule('/$model/show/<key>', endpoint=endpoints[OPE_SHOW]),
+    Rule('/$model/create', endpoint=endpoints[OPE_CREATE]),
+    Rule('/$model/update/<key>', endpoint=endpoints[OPE_UPDATE]),
+    Rule('/$model/delete/<key>', endpoint=endpoints[OPE_DELETE]),
   ])
 
   def __init__(self, model=None, **kwargs):
@@ -60,40 +78,47 @@ class CRUDViewGroup(ViewGroup):
     self.model_name_lower = self.model_name.lower()
 
   def get_additional_context_on_create(self, request, form):
-    return {}
+    if self.owner_attr:
+      if request.user.is_anonymous():
+        owner = None
+      else:
+        owner = request.user.key()
+      return {self.owner_attr: owner}
+    else:
+      return {}
 
   def get_additional_context_on_update(self, request, form):
     return {}
 
-  def get_query(self):
+  def get_query(self, request):
     if hasattr(self.model, 'created'):
       return self.model.all().order('-created')
     else:
       return self.model.all()
 
-  def get_template(self, name):
+  def get_template(self, request, name):
     return self.templates[name]
 
-  def get_form(self, name):
+  def get_form(self, request, name):
     try:
       return self.forms[name]
     except KeyError:
       return self.form
 
   def get_list_url(self, cursor=None):
-    return url_for(self.get_endpoint('list'), cursor=cursor)
+    return url_for(self.get_endpoint(OPE_LIST), cursor=cursor)
 
   def get_detail_url(self, obj):
-    return url_for(self.get_endpoint('show'), key=obj.key())
+    return url_for(self.get_endpoint(OPE_SHOW), key=obj.key())
 
   def get_delete_url(self, obj):
-    return url_for(self.get_endpoint('delete'), key=obj.key())
+    return url_for(self.get_endpoint(OPE_DELETE), key=obj.key())
 
   def get_update_url(self, obj):
-    return url_for(self.get_endpoint('update'), key=obj.key())
+    return url_for(self.get_endpoint(OPE_UPDATE), key=obj.key())
 
   def get_create_url(self):
-    return url_for(self.get_endpoint('create'))
+    return url_for(self.get_endpoint(OPE_CREATE))
 
   def url_processor(self, request):
     return {'list_url': self.get_list_url,
@@ -102,9 +127,30 @@ class CRUDViewGroup(ViewGroup):
             'update_url': self.get_update_url,
             'create_url': self.get_create_url}
 
+  def authorize(self, request, operation, obj=None):
+    """ Raise AuthorizationError when the operation is not permitted.
+    """
+    return True
+
+  def check_authority(self, request, operation, obj=None):
+    try:
+      self.authorize(request, operation, obj)
+    except NotAuthorized, e:
+      from kay.conf import settings
+      if 'kay.auth.middleware.AuthenticationMiddleware' in \
+            settings.MIDDLEWARE_CLASSES and \
+            request.user.is_anonymous():
+        from kay.utils import create_login_url
+        return redirect(create_login_url(request.url))
+      else:
+        raise Forbidden("Access not allowed.")
+
   def list(self, request, cursor=None):
     # TODO: bi-directional pagination instead of one way ticket forward
-    q = self.get_query()
+    ret = self.check_authority(request, OPE_LIST)
+    if ret:
+      return ret
+    q = self.get_query(request)
     if cursor:
       q = q.with_cursor(cursor)
     entities = q.fetch(self.entities_per_page)
@@ -112,7 +158,7 @@ class CRUDViewGroup(ViewGroup):
     q = q.with_cursor(next_cursor)
     if q.get() is None:
       next_cursor = None
-    return render_to_response(self.get_template('list'),
+    return render_to_response(self.get_template(request, OPE_LIST),
                               {'model': self.model_name,
                                'entities': entities,
                                'cursor': next_cursor,
@@ -129,7 +175,10 @@ class CRUDViewGroup(ViewGroup):
       entity = None
     if entity is None:
       raise NotFound("Specified %s not found." % self.model_name)
-    return render_to_response(self.get_template('show'),
+    ret = self.check_authority(request, OPE_SHOW, entity)
+    if ret:
+      return ret
+    return render_to_response(self.get_template(request, OPE_SHOW),
                               {'entity': entity,
                                'model': self.model_name},
                               processors=(self.url_processor,))
@@ -143,13 +192,19 @@ class CRUDViewGroup(ViewGroup):
         entity = None
       if entity is None:
         raise NotFound("Specified %s not found." % self.model_name)
-      form_class = self.get_form('update')
+      form_class = self.get_form(request, OPE_UPDATE)
       form = form_class(instance=entity)
       title = _("Updating a %s entity") % self.model_name
+      ret = self.check_authority(request, OPE_UPDATE, entity)
+      if ret:
+        return ret
     else:
-      form_class = self.get_form('create')
+      form_class = self.get_form(request, OPE_CREATE)
       form = form_class()
       title = _("Creating a new %s") % self.model_name
+      ret = self.check_authority(request, OPE_CREATE)
+      if ret:
+        return ret
     if request.method == 'POST':
       if form.validate(request.form, request.files):
         if key:
@@ -163,7 +218,7 @@ class CRUDViewGroup(ViewGroup):
         new_entity = form.save(**additional_context)
         set_flash(message)
         return redirect(self.get_list_url())
-    return render_to_response(self.get_template('update'),
+    return render_to_response(self.get_template(request, OPE_UPDATE),
                               {'form': form.as_widget(),
                                'title': title,
                                },
@@ -184,6 +239,9 @@ class CRUDViewGroup(ViewGroup):
       entity = None
     if entity is None:
       raise NotFound("Specified %s not found." % self.model_name)
+    ret = self.check_authority(request, OPE_DELETE, entity)
+    if ret:
+      return ret
     entity.delete()
     set_flash(_("An entity is deleted successfully."))
     # TODO: back to original page
